Ported the unit tests to work with the new permissions system

+++ kanbanik-server/src/main/scala/com/googlecode/kanbanik/commands/DeleteUserCommand.scala
@@ -13,13 +13,18 @@ class DeleteUserCommand extends Command[UserDto, EmptyDto] with CredentialsUtils
     if (user.unloggedFakeUser) {
       Right(ErrorDto("Can not delete the unloggded user."))
     } else {
-      // this is an expensive operation - consider to have this info on the user directly to speed things up
-      val tasksOfUser = for (board <- Board.all(includeTasks = true, User().withAllPermissions()); task <- board.tasks if task.assignee.getOrElse(User()).name == user.name) yield task
-      if (tasksOfUser.size == 0) {
-        user.copy(version = params.version).delete()
-        Left(EmptyDto())
+      val all = User.all(User().withAllPermissions())
+      if (all.filter(!_.unloggedFakeUser).size > 1) {
+        // this is an expensive operation - consider to have this info on the user directly to speed things up
+        val tasksOfUser = for (board <- Board.all(includeTasks = true, User().withAllPermissions()); task <- board.tasks if task.assignee.getOrElse(User()).name == user.name) yield task
+        if (tasksOfUser.size == 0) {
+          user.copy(version = params.version).delete()
+          Left(EmptyDto())
+        } else {
+          Right(ErrorDto("This user has tasks assigned - please delete them firs. Tasks: " + tasksOfUser.map(_.ticketId).mkString(", ")))
+        }
       } else {
-        Right(ErrorDto("This user has tasks assigned - please delete them firs. Tasks: " + tasksOfUser.map(_.ticketId).mkString(", ")))
+        Right(ErrorDto("You can not delete the last user from the system - it would not be possible to log in again!"))
       }
     }
   }
